Move tool definitions to separate files
